refactor(auth use-cases): update UseCase interface import
