+++ src/modules/auth/application/use-cases/get-current-session.ts
@@ -1,4 +1,4 @@
-import type { UseCase } from '@/shared/application/use-case';
+import type { UseCase } from '@/shared/application/shared.use-case';
 import type { IAuthRepository } from '@/modules/auth/domain/auth.respository';
 
 export class GetCurrentSession implements UseCase {
